Move field conditions
